Minor bug fixes in histograms and merge_sce function
The histograms were subsetting data to <= 2*stdev to focus on the majority of the data.  With ggplotly this isn't necessary (zoom).

merge_sce was using metadata which isn't exported.  Now using object@metadata for assignment.

Warnings on qc report temp dir creation now hidden (issue when generating multiple reports on a local machine).